% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reading_Helper_Functions.R
\name{AG_collapse}
\alias{AG_collapse}
\title{Collapse primary accelerometer data}
\usage{
AG_collapse(AG, output_window_secs, samp_freq)
}
\arguments{
\item{AG}{a dataframe of raw primary accelerometer data}

\item{output_window_secs}{the desired epoch length; defaults to one second}

\item{samp_freq}{The sampling frequency

raw_file <-
    system.file("extdata",
        "Raw AG to Collapse.csv",
        package = "TwoRegression")
raw <-
    read.csv(raw_file)
names(raw) <-
    gsub("\\.", " ", names(AG))AG_collapse(raw, 1, 80)
AG_collapse(raw, 1, 80)}
}
\description{
Collapse primary accelerometer data
}
\keyword{internal}
