% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing_Functions.R
\name{AG_smooth}
\alias{AG_smooth}
\title{Smooth predicted values from a two-regression algorithm}
\usage{
AG_smooth(AG, timestamps, epoch = 60, class_summary = FALSE, verbose)
}
\arguments{
\item{AG}{A data frame of monitor data that has been processed with a two-regression algorithm}

\item{timestamps}{A vector of POSIX-formatted timestamps corresponding to the two-regression data}

\item{epoch}{The desired epoch length over which to collapse}

\item{class_summary}{A logical scalar: Include summary of two-regression classification in the output?}

\item{verbose}{A logical scalar: print progress updates?}
}
\value{
A data frame of collapsed two-regression data
}
\description{
Collapses data over a specified epoch length
}
\examples{
\dontrun{
raw_file <-
    system.file("extdata",
        "TestID_LeftWrist_RAW.csv",
        package = "TwoRegression")

wear <- "Left Wrist"
id <- "Test"
alg <- 1:2
full_data <-
    hibbing18_twoReg_process(raw_file, imu_file, wear, id, alg)

AG_smooth(full_data, as.POSIXlt(full_data$Timestamp))
}
}
