% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_TwoRegressionProcessing_Helper_Functions.R
\name{algorithm_verify}
\alias{algorithm_verify}
\title{Identify mis-specified Hibbing 2018 algorithms, based on files provided}
\usage{
algorithm_verify(IMU, Algorithm)
}
\arguments{
\item{IMU}{A character scalar giving path to IMU data file}

\item{Algorithm}{A numeric vector giving the algorithm(s) to apply to the data from the primary accelerometer and (if applicable) IMU}
}
\description{
Identify mis-specified Hibbing 2018 algorithms, based on files provided
}
\examples{
\dontrun{
algorithm_verify(NULL, 4)
algorithm_verify(NULL, NULL)
}

}
\keyword{internal}
