% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reading_Helper_Functions.R
\name{check_columns}
\alias{check_columns}
\title{Check if the primary accelerometer file is formatted correctly}
\usage{
check_columns(file)
}
\arguments{
\item{file}{A character scalar giving path to primary accelerometer file}
}
\description{
Check if the primary accelerometer file is formatted correctly
}
\keyword{internal}
