% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{classify_validation_data}
\alias{classify_validation_data}
\title{Establish an annotated validation data set for
leave-one-participant-out-cross-validation}
\usage{
classify_validation_data(data, sed_cp = 0, ambulation_cp = 0,
  sed_cpvar = "ENMO", ambulation_cpvar = "ENMO_meancv")
}
\arguments{
\item{data}{The data set to annotate}
}
\description{
Establish an annotated validation data set for
leave-one-participant-out-cross-validation
}
\keyword{internal}

