% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crouter__main.R
\name{crouter}
\alias{crouter}
\alias{crouter_2006}
\alias{crouter_2010}
\alias{crouter_2012}
\title{Apply two-regression methods from publications by Crouter et al.}
\usage{
crouter_2006(AG, movement_var = "Axis1", time_var = "Timestamp")

crouter_2010(AG, movement_var = "Axis1", time_var = "Timestamp")

crouter_2012(
  AG,
  movement_var = "Axis1",
  time_var = "Timestamp",
  model = c("VA", "VM"),
  check = TRUE
)
}
\arguments{
\item{AG}{data frame containing ActiGraph data (activity counts)}

\item{movement_var}{character scalar. Name of the movement variable (default
is \code{"Axis1"})}

\item{time_var}{character scalar. Name of the timestamp variable (required to
verify that input epoch length is 10 seconds)}

\item{model}{character scalar. One of \code{"VA"} (for the vertical axis,
youth-specific model of Crouter et al., 2012) or \code{"VM"} (for the
vector magnitude youth-specific model from the same paper)}

\item{check}{Logical. Should a warning be issued that will prompt you to
check that the selected model matches the selected  movement variable?}
}
\value{
The original data appended with columns giving activity
  classification (sedentary, ambulatory, or intermittent) and energy
  expenditure (i.e, METs)
}
\description{
Apply two-regression methods from publications by Crouter et al.
}
\examples{
data(count_data, package = "TwoRegression")
crouter_2006(count_data, "Axis1", "time")
crouter_2010(count_data, "Axis1", "time")
crouter_2012(count_data, "Axis1", "time", "VA", FALSE)
crouter_2012(count_data, "Vector.Magnitude", "time", "VM", FALSE)

}
\seealso{
\href{https://pubmed.ncbi.nlm.nih.gov/16322367/}{Crouter et al. (2006, \emph{J Appl Physiol})}
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2891855/}{Crouter et al. (2010, \emph{Med Sci Sports Exerc})}
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3324667/}{Crouter et al. (2012, \emph{Med Sci Sports Exerc})}
}
