% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_ActiGraph_Scripts.R
\name{cv}
\alias{cv}
\title{Calculate coefficient of variation}
\usage{
cv(signal)
}
\arguments{
\item{signal}{the variable on which to perform calculation}
}
\description{
Calculate coefficient of variation
}
\keyword{internal}
