% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hibbing_directions.R
\name{directions}
\alias{directions}
\alias{get_directions}
\alias{get_dcp5}
\title{Calculate direction changes per five seconds in sliding windows}
\usage{
get_directions(x, window_size = 5)

get_dcp5(x, window_size = 5)
}
\arguments{
\item{x}{a numeric vector on which to perform the calculation}

\item{window_size}{width of the sliding window, in data points}
}
\value{
a numeric vector of values, giving the number of direction
  changes in the sliding window that corresponds to each epoch of data
}
\description{
\code{get_directions} is legacy code, whereas \code{get_dcp5} is cleaner and
(possibly) faster
}
\keyword{internal}
