% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_ActiGraph_Scripts.R
\name{get.directions}
\alias{get.directions}
\title{Calculate direction changes per five seconds in sliding windows}
\usage{
get.directions(big_data, window_secs = 5)
}
\arguments{
\item{big_data}{a numeric vector on which to perform the calculation}

\item{window_secs}{size of the sliding window, in seconds}
}
\value{
a numeric vector of values, giving the number of direction changes in the sliding window that corresponds to each epoch of data
}
\description{
Calculate direction changes per five seconds in sliding windows
}
