% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates&Times.R
\name{get.excel.time}
\alias{get.excel.time}
\title{Convert Excel Times.}
\usage{
get.excel.time(numericTime, tz = "GMT")
}
\arguments{
\item{numericTime}{The Excel serial time}

\item{tz}{The desired timezone}
}
\description{
Converts serial times from Excel into POSIX format in GMT.
}
\keyword{internal}
