% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cv_static.R
\name{get_cv_static}
\alias{get_cv_static}
\title{Calculate coefficient of variation for groupings of accelerometer data points}
\usage{
get_cv_static(big_data, window_size = 10, verbose = FALSE)
}
\arguments{
\item{big_data}{a numeric vector on which to perform the calculation}

\item{window_size}{width of the sliding window, in data points}

\item{verbose}{logical. Print progress updates?}
}
\value{
a numeric vector of coefficient of variation values
}
\description{
Calculate coefficient of variation for groupings of accelerometer data points
}
\keyword{internal}
