% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_ActiGraph_Scripts.R
\name{get_cv_vars}
\alias{get_cv_vars}
\title{Determine which variable(s) to calculate coefficient of variation on}
\usage{
get_cv_vars(Algorithm, verbose = FALSE)
}
\arguments{
\item{Algorithm}{the algorithm(s) selected to process the accelerometer/IMU data}
}
\description{
Determine which variable(s) to calculate coefficient of variation on
}
\keyword{internal}
