% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates&Times.R
\name{get_day_of_year-deprecated}
\alias{get_day_of_year-deprecated}
\title{Julian Date}
\usage{
get_day_of_year(timestamp, format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\value{
A numeric vector of Julian dates.
}
\description{
A wrapper to retrieve the Julian date.
}
\examples{
\dontrun{
key_dates <- c("2018-01-01", "2018-12-31")

# Warns about deprecation
TwoRegression:::get_day_of_year(key_dates, "\%Y-\%m-\%d")

# Use the following instead
AGread::get_day_of_year(key_dates, "\%Y-\%m-\%d")
}

}
\seealso{
\code{\link{TwoRegression-deprecated}}
}
\keyword{internal}
