% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_ActiGraph_Scripts.R
\name{get_file_meta}
\alias{get_file_meta}
\title{Get file metadata (sampling frequency and timestamps) for primary accelerometer}
\usage{
get_file_meta(file)
}
\arguments{
\item{file}{character scalar giving path to primary accelerometer file}
}
\description{
Get file metadata (sampling frequency and timestamps) for primary accelerometer
}
\keyword{internal}
