% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Functions.R
\name{getVM}
\alias{getVM}
\title{Calculate vector magnitude}
\usage{
getVM(triaxial, verbose = T)
}
\arguments{
\item{triaxial}{a dataframe of triaxial data on which to calculate vector magnitude}

\item{verbose}{print information about variable search criteria?}
}
\value{
a vector of vector magnitude values
}
\description{
Calculate vector magnitude
}
