% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hibbing__main.R
\name{hibbing_2018}
\alias{hibbing_2018}
\title{Apply two-regression models from Hibbing et al.}
\usage{
hibbing_2018(
  AG,
  accel_var = "ENMO",
  gyro_var = "GVM",
  direction_var = "Direction",
  time_var = "Timestamp",
  site = c("Hip", "Left Wrist", "Right Wrist", "Left Ankle", "Right Ankle"),
  algorithm = 1:3,
  smooth = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{AG}{data frame containing ActiGraph data (activity counts)}

\item{accel_var}{Character scalar. Name of accelerometer variable to
operate on (expected format is Euclidian Norm Minus One, in
milli-gravitational units)}

\item{gyro_var}{character scalar. Name of gyroscope variable to operate
on (expected format is gyroscope vector magnitude, in degrees per second)}

\item{direction_var}{character scalar. Name of magnetometer variable to
operate on (expected format is a vector of directions, likely produced by
\code{AGread::classify_magnetometer})}

\item{time_var}{character scalar. Name of the timestamp variable (required to
verify that input epoch length is 10 seconds)}

\item{site}{character scalar or vector of attachment sites (more than one may
be desired, e.g., if results are to be compared from running both of the
wrist-specific algorithms)}

\item{algorithm}{An integer/numeric scalar or vector giving the algorithm(s)
to apply to the data from the primary accelerometer and (if applicable)
IMU. Must be \code{1} (accelerometer only), \code{2} (accelerometer and
gyroscope), \code{3} (accelerometer, gyroscope, and magnetometer), or any
combination thereof}

\item{smooth}{A logical scalar. Should data be averaged over a longer time period after processing?}

\item{verbose}{lobical. Print progress updates?}

\item{...}{Further arguments passed to \code{\link{hibbing_smooth}}}
}
\description{
Apply two-regression models from Hibbing et al.
}
\examples{
data(all_data, package = "TwoRegression")

result <- hibbing_2018(
  all_data, "ENMO", "Gyroscope_VM_DegPerS", "mean_magnetometer_direction",
  site = "Hip", algorithm = 1
)
utils::head(result)

collapsed <- hibbing_2018(
  all_data, "ENMO", "Gyroscope_VM_DegPerS", "mean_magnetometer_direction",
  site = "Hip", algorithm = 1:2, smooth = TRUE
)
collapsed

}
\seealso{
\href{https://pubmed.ncbi.nlm.nih.gov/29271847/}{Hibbing et al. (2018,
\emph{Med Sci Sports Exerc})}

\code{\link{apply_two_regression_hibbing18}}
\code{\link{hibbing_smoot}}
}
