% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reading_Helper_Functions.R
\name{imu_collapse}
\alias{imu_collapse}
\title{Collapse raw IMU data to a specified epoch}
\usage{
imu_collapse(AG, block_size, verbose = FALSE)
}
\arguments{
\item{AG}{dataframe containing raw IMU data}

\item{block_size}{number of samples per epoch}

\item{verbose}{A logical scalar: print progress updates?}
}
\value{
dataframe of IMU data averaged over the specified epoch length
}
\description{
Collapse raw IMU data to a specified epoch
}
\examples{
imu_file <-
    system.file("extdata",
        "IMU Data to Collapse.csv",
        package = "TwoRegression")

imu <- read.csv(imu_file)

imu_collapse(imu, 100)

}
\keyword{internal}
