% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Functions.R
\name{imu_filter_gyroscope-deprecated}
\alias{imu_filter_gyroscope-deprecated}
\title{Low-Pass filter the Gyroscope data at 35 Hz}
\usage{
imu_filter_gyroscope(AG, samp_rate, verbose = FALSE)
}
\value{
a vector of vector magnitude values
}
\description{
Low-Pass filter the Gyroscope data at 35 Hz
}
\examples{
\dontrun{
data(imu_to_collapse)
TwoRegression:::imu_filter_gyroscope(imu)
}

}
\seealso{
\code{\link{TwoRegression-deprecated}}
}
\keyword{internal}
