% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c_data_doc.R
\docType{data}
\name{imu_to_check}
\alias{imu_to_check}
\title{IMU data to check}
\format{
A data frame with 300 rows and 8 variables:
\describe{
  \item{file_source_IMU}{The filename of the IMU file}
  \item{date_processed_IMU}{The date the IMU file was processed}
  \item{Timestamp}{The corresponding time for each row of data}
  \item{Gyroscope_VM_DegPerS}{Gyroscope vector magnitude, in degrees per
  second}
  \item{mean_abs_Gyroscope_x_DegPerS}{Rotation in x axis, degrees per second}
  \item{mean_abs_Gyroscope_y_DegPerS}{Rotation in y axis, degrees per second}
  \item{mean_abs_Gyroscope_z_DegPerS}{Rotation in z axis, degrees per second}
  \item{mean_magnetometer_direction}{Cardinal direction of magnetometer
  signal, averaged over one second}
}
}
\usage{
imu_to_check
}
\description{
A dataset for demonstrating checks that are applied to IMU data.
}
\keyword{datasets}
