% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c_data_doc.R
\docType{data}
\name{imu_to_collapse}
\alias{imu_to_collapse}
\title{IMU data to collapse}
\format{
A data frame with 1500 rows and 17 variables:
\describe{
  \item{Timestamp}{The corresponding time for each row of data}
  \item{Accelerometer.X}{Secondary accelerometer x-axis data, in G}
  \item{Accelerometer.Y}{Secondary accelerometer y-axis data, in G}
  \item{Accelerometer.Z}{Secondary accelerometer z-axis data, in G}
  \item{Temperature}{Temperature of the IMU, in Celcius}
  \item{Gyroscope.X}{Gyroscope x-axis data, in degrees per second}
  \item{Gyroscope.Y}{Gyroscope y-axis data, in degrees per second}
  \item{Gyroscope.Z}{Gyroscope z-axis data, in degrees per second}
  \item{Magnetometer.X}{Magnetometer x-axis data, in micro-Teslas}
  \item{Magnetometer.Y}{Magnetometer y-axis data, in micro-Teslas}
  \item{Magnetometer.Z}{Magnetometer z-axis data, in micro-Teslas}
  \item{file_source_IMU}{The filename of the IMU file}
  \item{date_processed_IMU}{The date the IMU file was processed}
  \item{ms}{The millisecond value of the timestamp}
  \item{mean_Accel_VM}{Vector magnitude of the secondary accelerometer
  signal, in G}
  \item{Gyroscope_VM_DegPerS}{Gyroscope vector magnitude, in degrees per
  second}
  \item{Magnetometer_VM_MicroT}{Vector magnitude of the magnetometer signal,
  in micro-Teslas}
}
}
\usage{
imu_to_collapse
}
\description{
A partially-processed IMU dataset ready to be collapsed from raw samples to
one-second summaries.
}
\keyword{datasets}
