% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2rm_helpers.R
\name{plot_sed}
\alias{plot_sed}
\title{Plot behaviors for visualizing a sedentary cut-point}
\usage{
plot_sed(
  object,
  sed_cp_activities,
  sed_activities,
  activity_var,
  sed_cpVar,
  met_var,
  ...
)
}
\arguments{
\item{object}{TwoRegression object to plot}

\item{sed_cp_activities}{activities to include in a subset of \code{AG}}

\item{sed_activities}{Character vector. Actual sedentary activities}

\item{activity_var}{Character scalar. Name of the variable defining which
activity is being performed}

\item{sed_cpVar}{character scalar. The name of the variable on which the
cut-point is based}

\item{met_var}{character scalar. The name of the variable containing energy
expenditure values, in metabolic equivalents}

\item{...}{additional arguments passed to \code{\link{plotCP}}}
}
\description{
Plot behaviors for visualizing a sedentary cut-point
}
\keyword{internal}
