% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2rm_helpers.R
\name{plot_walkrun}
\alias{plot_walkrun}
\title{Plot behaviors for visualizing a walk/run cut-point}
\usage{
plot_walkrun(
  object,
  activity_var,
  walkrun_activities,
  walkrun_cpVar,
  met_var,
  ...
)
}
\arguments{
\item{object}{the \code{TwoRegression} object to plot from}

\item{activity_var}{Character scalar. Name of the variable defining which
activity is being performed}

\item{walkrun_activities}{character vector giving values of
\code{activity_var} that are walk/run activities}

\item{walkrun_cpVar}{character scalar giving the name of the variable on
which the walk/run cut-point is based}

\item{met_var}{character scalar. The name of the variable containing energy
expenditure values, in metabolic equivalents}

\item{...}{further arguments passed to plotting calls}
}
\description{
Plot behaviors for visualizing a walk/run cut-point
}
\keyword{internal}
