% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_ActiGraph_Scripts.R
\name{read.AG.raw}
\alias{read.AG.raw}
\title{File reading function for primary accelerometer files}
\usage{
read.AG.raw(file, output.window = 1)
}
\arguments{
\item{file}{A character scalar giving path to primary accelerometer file}

\item{output.window}{The desired epoch length; defaults to one second}
}
\value{
A dataframe giving processed raw data from the primary accelerometer in the specified epoch length
}
\description{
File reading function for primary accelerometer files
}
