% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reading_Functions.R
\name{read_AG_raw-deprecated}
\alias{read_AG_raw-deprecated}
\title{File reading function for primary accelerometer files}
\usage{
read_AG_raw(file, output_window_secs = 1, verbose = FALSE)
}
\arguments{
\item{file}{A character scalar giving path to primary accelerometer file}

\item{output_window_secs}{the desired epoch length; defaults to one second}
}
\value{
A dataframe giving processed raw data from the primary accelerometer in the specified epoch length
}
\description{
File reading function for primary accelerometer files
}
\examples{
raw_file <-
    system.file("extdata",
    "TestID_LeftWrist_RAW.csv",
    package = "TwoRegression")

\dontrun{
# Gives warning about deprecation
  read_AG_raw(raw_file)

# Use the following instead
  AGread::read_AG_raw(raw_file)
}
}
\seealso{
\code{\link{TwoRegression-deprecated}}
}
\keyword{internal}
