% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing_Functions.R
\name{smooth_column}
\alias{smooth_column}
\title{Rules to collapse columns of data passed to \code{\link{AG_smooth}}}
\usage{
smooth_column(col_name, AG)
}
\arguments{
\item{col_name}{The name of a single column in \code{AG} to be collapsed}

\item{AG}{A data frame of monitor data that has been processed with a two-regression algorithm}
}
\value{
A vector of data collapsed according to the specified rule for \code{col_name}
}
\description{
Rules to collapse columns of data passed to \code{\link{AG_smooth}}
}
\keyword{internal}
