% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates&Times.R, R/File_Reading_Functions.R,
%   R/TwoRegression-deprecated.R
\name{get_minute}
\alias{get_minute}
\alias{get_day_of_year}
\alias{read_AG_raw}
\alias{read_IMU}
\alias{TwoRegression-deprecated}
\title{Deprecated functions in package \pkg{TwoRegression}.}
\usage{
get_minute(timestamp, format = "\%Y-\%m-\%d \%H:\%M:\%S", rational = FALSE)

get_day_of_year(timestamp, format = "\%Y-\%m-\%d \%H:\%M:\%S")

read_AG_raw(file, output_window_secs = 1, verbose = FALSE)

read_IMU(file, output_window_secs = 1, verbose = FALSE)
}
\description{
These are functions that are deprecated because they have been
  moved elsewhere, primarily to the AGread package. Use
  \code{help("-deprecated")} to access documentation for deprecated
  functions.
}
\section{\code{get_minute}}{

For \code{get_minute}, use \code{\link[AGread]{get_minute}}
}

\section{\code{get_day_of_year}}{

For \code{get_day_of_year} use \code{\link[AGread]{get_day_of_year}}
}

\section{\code{read_AG_raw}}{

For \code{read_AG_raw}, use \code{\link[AGread]{read_AG_raw}}
}

\section{\code{read_IMU}}{

For \code{read_IMU}, use \code{\link[AGread]{read_AG_IMU}}
}

\keyword{internal}
