% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoRegression.R
\name{TwoRegression-Function}
\alias{TwoRegression-Function}
\alias{TwoRegression}
\title{Wrapper for applying two-regression models}
\usage{
TwoRegression(
  AG,
  method = c("Hibbing_2018", "Crouter_2012", "Crouter_2010", "Crouter_2006"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{AG}{data frame of actigraph data}

\item{method}{character object (scalar or vector) telling which model(s) to
apply to the data}

\item{verbose}{logical. Print updates to console?}

\item{...}{}
}
\description{
Wrapper for applying two-regression models
}
\examples{

data(all_data, package = "TwoRegression")

TwoRegression(
  all_data, "Hibbing_2018", verbose = FALSE,
  site = c("Left Wrist", "Right Wrist"),
  algorithm = 1:2, accel_var = "ENMO",
  gyro_var = "Gyroscope_VM_DegPerS", smooth = FALSE
)
}
