% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0b_package_doc.R
\docType{package}
\name{TwoRegression-Package}
\alias{TwoRegression-Package}
\title{Develop and Apply Two-Regression Algorithms}
\description{
The TwoRegression package is designed to make working with
two-regression algorithms quick, easy, and accurate.
}
\details{
Originally, the package was designed to house the algorithms created by
Hibbing et al. (2018). Since then, support has been added for other
algorithms, including Crouter et al. (2006), Crouter et al. (2010), and
Crouter et al. (2012). Functionality has also been added to develop and
cross-validate new two-regression algorithms. The package \code{RcppRoll}
has also been invoked to speed up rolling coefficient of variation
calculations.
}
\section{Associated References}{


Hibbing PR, LaMunion SR, Kaplan AS, & Crouter SE (2018). Estimating
energy expenditure with ActiGraph GT9X Inertial Measurement Unit.
\emph{Medicine and Science in Sports and Exercise}. 50(5), 1093-1102.
doi: 10.1249/MSS.0000000000001532

Crouter, S. E., Clowers, K. G., & Bassett Jr, D. R. (2006). A novel method
for using accelerometer data to predict energy expenditure. \emph{Journal of
Applied Physiology}, 100(4), 1324-1331.

Crouter, S. E., Kuffel, E., Haas, J. D., Frongillo, E. A., & Bassett Jr, D.
R. (2010). Refined Two-Regression Model for the ActiGraph Accelerometer.
\emph{Medicine and Science in Sports and Exercise}, 42(5), 1029.

Crouter, S. E., Horton, M., & Bassett Jr, D. R. (2012). Use of a
Two-regression model for estimating energy expenditure in children.
\emph{Medicine and Science in Sports and Exercise}, 44(6), 1177.
}

\examples{
\donttest{

data(count_data, package = "TwoRegression")
crouter_2006(count_data, "Axis1", "time")
crouter_2010(count_data, "Axis1", "time")
crouter_2012(count_data, "Axis1", "time", "VA", FALSE)
crouter_2012(count_data, "Vector.Magnitude", "time", "VM", FALSE)

}


}
