% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoRegression.R
\docType{package}
\name{TwoRegression}
\alias{TwoRegression}
\alias{TwoRegression-package}
\title{Process Accelerometer Data with Two-Regression Algorithms}
\description{
The critical functions of the TwoRegression package are:
}
\section{Core functions}{

}

\section{Associated References}{

Hibbing PR, LaMunion SR, Kaplan AS, & Crouter SE (2017). Estimating
energy expenditure with ActiGraph GT9X Inertial Measurement Unit.
*Medicine and Science in Sports and Exercise*. Advance online publication.
doi:
}

