% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_ActiGraph_Scripts.R
\name{AG_collapse}
\alias{AG_collapse}
\title{Collapse primary accelerometer data}
\usage{
AG_collapse(AG, output.window, samp.freq)
}
\arguments{
\item{AG}{a dataframe of raw primary accelerometer data}
}
\description{
Collapse primary accelerometer data
}
\keyword{internal}
