% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reading_Helper_Functions.R
\name{AG_collapse}
\alias{AG_collapse}
\title{Collapse primary accelerometer data}
\usage{
AG_collapse(AG, output_window, samp_freq)
}
\arguments{
\item{AG}{a dataframe of raw primary accelerometer data}
}
\description{
Collapse primary accelerometer data
}
\keyword{internal}
