% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_smooth.R
\name{ag_smooth}
\alias{ag_smooth}
\title{Smooth two-regression estimates over specified periods}
\usage{
ag_smooth(AG, time_var = "Timestamp", unit = "60 sec", verbose = FALSE)
}
\arguments{
\item{AG}{data frame containing ActiGraph data (activity counts)}

\item{time_var}{character scalar. Name of the timestamp variable (required to
verify that input epoch length is 10 seconds)}

\item{unit}{the interval to use for smoothing (see
\code{\link[lubridate]{floor_date}}). Default is \code{"60 sec"}}

\item{verbose}{lobical. Print progress updates?}
}
\value{
Smoothed data, collapsed in the specified intervals
}
\description{
Smooth two-regression estimates over specified periods
}
\keyword{internal}
