% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c_data_doc.R
\docType{data}
\name{all_data}
\alias{all_data}
\title{Two-regression-ready data frame}
\format{
A data frame with 299 rows and 17 variables:
\describe{
  \item{PID}{Participant ID}
  \item{file_source_PrimaryAccel}{The filename of the primary accelerometer
  file}
  \item{date_processed_PrimaryAccel}{The date the primary accelerometer file
  was processed}
  \item{file_source_IMU}{The filename of the IMU file}
  \item{date_processed_IMU}{The date the IMU file was processed}
  \item{Timestamp}{The corresponding time for each row of data}
  \item{day_of_year}{The numeric day of the year, i.e., the Julian date}
  \item{minute_of_day}{The numeric minute of the day}
  \item{ENMO}{Euclidian Norm Minus One, in milli-g}
  \item{Gyroscope_VM_DegPerS}{Gyroscope vector magnitude, in degrees per
  second}
  \item{mean_abs_Gyroscope_x_DegPerS}{Rotation in x axis, degrees per second}
  \item{mean_abs_Gyroscope_y_DegPerS}{Rotation in y axis, degrees per second}
  \item{mean_abs_Gyroscope_z_DegPerS}{Rotation in z axis, degrees per second}
  \item{mean_magnetometer_direction}{Cardinal direction of magnetometer
  signal, averaged over one second}
  \item{ENMO_CV10s}{Coefficient of variation per 10-s, applied to Euclidian
  Norm Minus One}
  \item{GVM_CV10s}{Coefficient of variation per 10-s, applied to gyroscope
  vector magnitude}
  \item{Direction}{Direction changes per 5-s}
}
}
\usage{
all_data
}
\description{
A dataset with pre-processed primary accelerometer and IMU data that is ready
for applying a two-regression algorithm.
}
\keyword{datasets}
