% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_TwoRegressionProcessing_Helper_Functions.R
\name{attachment_verify}
\alias{attachment_verify}
\title{Identify mis-specified Hibbing 2018 attachment site}
\usage{
attachment_verify(Wear_Location)
}
\arguments{
\item{Wear_Location}{A character scalar indicating the device's attachment site}
}
\description{
Identify mis-specified Hibbing 2018 attachment site
}
\examples{
\dontrun{
attachment_verify("Left Wrist")
attachment_verify(c("Left Wirst", "Right Wrist"))
attachment_verify("Left Wirst")
}

}
\keyword{internal}
