% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reading_Helper_Functions.R
\name{check_second}
\alias{check_second}
\title{Check if the IMU data start on an exact second}
\usage{
check_second(AG)
}
\arguments{
\item{AG}{a dataframe of IMU data}
}
\description{
Check if the IMU data start on an exact second
}
\examples{
\dontrun{
imu_file <-
    system.file("extdata",
    "TestID_LeftWrist_IMU.csv",
    package = "TwoRegression")

AG <- read_IMU(imu_file)

check_second(AG)
}

}
\keyword{internal}
