% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c_data_doc.R
\docType{data}
\name{count_data}
\alias{count_data}
\title{Activity count data for demonstrating prior two-regression models}
\format{
A data frame with 30 rows and 5 variables:
\describe{
  \item{time}{POSIX. The timestamp}
  \item{Axis1}{numeric. The vertical axis activity counts}
  \item{Axis2}{numeric. The horizontal axis}
  \item{Axis3}{numeric. The lateral axis}
  \item{Vector.Magnitude}{numeric. The vector magnitude of all three axes}
}
}
\usage{
count_data
}
\description{
A small amount of 10-s epoch activity counts for code examples
}
\keyword{datasets}
