% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hibbing_directions.R
\name{directions}
\alias{directions}
\alias{get_directions}
\alias{get_dcp5}
\title{Calculate direction changes per five seconds in sliding windows}
\usage{
get_directions(x, window_size = 5)

get_dcp5(x, window_size = 5)
}
\arguments{
\item{x}{a numeric vector on which to perform the calculation}

\item{window_size}{width of the sliding window, in data points}
}
\value{
a numeric vector of values, giving the number of direction
  changes in the sliding window that corresponds to each epoch of data
}
\description{
\code{get_directions} is legacy code, whereas \code{get_dcp5} is cleaner and
(possibly) faster
}
\examples{
\donttest{
##All possible directions
directions <-
  c("N", "NNE", "NE", "ENE",
    "E", "ESE", "SE", "SSE",
    "S", "SSW", "SW", "WSW",
    "W", "WNW", "NW", "NNW")

##Reproducible results
set.seed(55)
direction_vector <- sample(directions, 50, replace = TRUE)

##Vector of direction changes per 5-s. First and last two values are always NA
get_directions(direction_vector)
get_dcp5(direction_vector)
}

}
\keyword{internal}
