% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{DualCP_LOSO}
\alias{DualCP_LOSO}
\title{Perform leave-one-participant-out-cross-validation on a two-regression
algorithm}
\usage{
DualCP_LOSO(model, subject_var = "id", data, MET_var = "MET_RMR",
  activity_var = "Behavior", verbose = FALSE, trace = FALSE)
}
\arguments{
\item{model}{A \code{TwoRegression} object formed with \code{\link{form_2rm}}
on which to perform the cross-validation}

\item{subject_var}{The variable that distinguishes between participants
(i.e., the fold-defining variable)}

\item{data}{The full data set to cross-validate}

\item{MET_var}{The outcome variable (in metabolic equivalents)}

\item{activity_var}{The activity being performed}

\item{verbose}{Logical. Print updates?}

\item{trace}{Logical. Print information about each iteration?}
}
\value{
A data frame with predictions obtained from
  leave-one-participant-out-cross-validation
}
\description{
Perform leave-one-participant-out-cross-validation on a two-regression
algorithm
}
\note{
This function will not work for \code{TwoRegression} objects formed
  from previously-published research. The \code{TwoRegression} object needs
  to have more information than is avaintermittentble in those cases in order to
  perform cross-validation, and this is sensible, since there is no reason or
  way to re-perform cross-validation on an already-finalized algorithm.
}
\keyword{internal}
