% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Form_2RM.R
\name{form_2rm}
\alias{form_2rm}
\alias{is.TwoRegression}
\title{Develop a two-regression algorithm}
\usage{
form_2rm(data, activity_var, sed_cp_activities, sed_activities, sed_cp_var,
  sed_METs, walkrun_activities, walkrun_cp_var, met_var, walkrun_formula,
  intermittent_formula)

is.TwoRegression(x)
}
\arguments{
\item{data}{The data with which to develop the algorithm}

\item{activity_var}{Character scalar. Name of the variable defining which
activity is being performed}

\item{sed_cp_activities}{Character vector. Activities to be included in the
process of forming the sedentary classifier}

\item{sed_activities}{Character vector. Actual sedentary activities}

\item{sed_cp_var}{Character scalar. Name of the variable on which the
sedentary cut-point is defined}

\item{sed_METs}{Numeric scalar. Metabolic equivalent value to apply to
sedentary activities}

\item{walkrun_activities}{Character vector. Actual ambulatory activities}

\item{walkrun_cp_var}{Character scalar. Name of the variable on which the
walk/run cut-point is defined}

\item{met_var}{Character scalar. Name of the variable giving actual energy
expenditure (in metabolic equivalents)}

\item{walkrun_formula}{Character scalar. Formula to use for developing the
walk/run regression model}

\item{intermittent_formula}{Character scalar. Formula to use for developing the
intermittent activity regression model}

\item{x}{object to be tested}
}
\value{
An object of class `TwoRegression`
}
\description{
Develop a two-regression algorithm

Check if an object has class TwoRegression
}
\examples{
\dontrun{
data(ag_metabolic_s1, package = "FLPAYr")
test_data <- subset(ag_metabolic_s1, site == "hip")
form_2rm(
data = test_data,
activity_var = "Behavior",
sed_cp_activities = c("Internet", "Reclining",
"Sweep", "Book", "Games", "Lying", "Dust"),
sed_activities = c("Internet", "Reclining", "Book", "Games", "Lying"),
sed_cp_var = "ENMO",
sed_METs = 1.25,
walkrun_activities = c("Run", "Walk_Slow", "Walk_Brisk"),
walkrun_cp_var = "ENMO_CV10s",
met_var = "MET_RMR",
walkrun_formula = "MET_RMR ~ ENMO",
intermittent_formula = "MET_RMR ~ I(ENMO)+I(ENMO^2)+I(ENMO^3)"
)
}
}
\seealso{
Links to methods for `TwoRegression` should go here (predict, plot,
  summary)
}
