% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_2rm_helpers.R
\name{get_2rm_formula}
\alias{get_2rm_formula}
\title{Convert an object of class TwoRegression to a textual representation of the
algorithm}
\usage{
get_2rm_formula(object)
}
\arguments{
\item{object}{the TwoRegression object}
}
\description{
Convert an object of class TwoRegression to a textual representation of the
algorithm
}
\keyword{internal}
