% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_2rm__main.R
\name{get_cut_point}
\alias{get_cut_point}
\title{Develop a cut-point as part of the process for developing a two-regression
algorithm}
\usage{
get_cut_point(
  data,
  activity_var,
  included_activities,
  positive_activities,
  cp_var
)
}
\arguments{
\item{data}{the data on which to perform the operation}

\item{activity_var}{character scalar. Name of the variable that indicates
what activity each data point corresponds to.}

\item{included_activities}{character vector. All activities included in the
data subset to be used for developing the cut-point}

\item{positive_activities}{character vector. Activities considered part of
the positive class}

\item{cp_var}{Character scalar. Name of the variable on which the cut-point
will be based.}
}
\value{
an object of class `roc`
}
\description{
Develop a cut-point as part of the process for developing a two-regression
algorithm
}
\keyword{internal}
