% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_ActiGraph_Scripts.R
\name{get_cv_vars}
\alias{get_cv_vars}
\title{Determine which variable(s) to calculate coefficient of variation on}
\usage{
get_cv_vars(Algorithm, verbose = FALSE)
}
\arguments{
\item{Algorithm}{the algorithm(s) selected to process the accelerometer/IMU
data}
}
\description{
\code{get_cv_vars} returns the name(s) of variables on which to calculate the
coefficient of variation in sliding windows, depending on which
two-regression algorithm(s) have been designated for application to the data.
}
\examples{
get_cv_vars(1)
get_cv_vars(2)
get_cv_vars(c(1,2))

}
\keyword{internal}
