% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_2rm_helpers.R
\name{get_lm_formula}
\alias{get_lm_formula}
\title{Convert a linear model to a text representation of the prediction equation}
\usage{
get_lm_formula(model)
}
\arguments{
\item{model}{The linear model object}
}
\description{
Convert a linear model to a text representation of the prediction equation
}
\keyword{internal}
