% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_TwoRegressionProcessing_MasterFunction.R
\name{hibbing18_twoReg_process}
\alias{hibbing18_twoReg_process}
\title{Process GT9X Files with Hibbing Two-Regression Algorithms}
\usage{
hibbing18_twoReg_process(RAW, IMU = NULL, Wear_Location = c("Hip",
  "Left Wrist", "Right Wrist", "Left Ankle", "Right Ankle"), PID,
  Algorithm = 1)
}
\arguments{
\item{RAW}{A character scalar giving path to primary accelerometer data file}

\item{IMU}{A character scalar giving path to IMU data file}

\item{Wear_Location}{A character scalar indicating the device's attachment site}

\item{PID}{A character scalar giving the participant identification}

\item{Algorithm}{A numeric vector giving the algorithm(s) to apply to the data from the primary accelerometer and (if applicable) IMU}
}
\value{
A data frame giving the data and predictions
}
\description{
Process GT9X Files with Hibbing Two-Regression Algorithms
}
