% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor_Functions.R
\name{imu_filter_gyroscope}
\alias{imu_filter_gyroscope}
\title{Low-Pass filter the Gyroscope data at 35 Hz}
\usage{
imu_filter_gyroscope(AG, samp_rate, verbose = FALSE)
}
\arguments{
\item{AG}{a dataframe of IMU data}
}
\description{
Low-Pass filter the Gyroscope data at 35 Hz
}
\examples{
imu_file <-
    system.file("extdata",
        "IMU Data to Collapse.csv",
        package = "TwoRegression")

imu <-
    read.csv(imu_file)

imu_filter_gyroscope(imu)

}
\keyword{internal}
