% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{message_update}
\alias{message_update}
\title{Print updates to console}
\usage{
message_update(message_number, vm_variables, duration, is_message = FALSE)
}
\arguments{
\item{message_number}{The index of the message to print}

\item{is_Message}{print update as a message?}
}
\description{
Print updates to console
}
\keyword{internal}
