% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{message_update}
\alias{message_update}
\title{Print updates to console}
\usage{
message_update(message_number, file, vm_variables, duration, cvs, window_size,
  is_message = FALSE, n)
}
\arguments{
\item{message_number}{the index of the message to print}

\item{file}{the name of the file being processed}

\item{vm_variables}{variables being used to calculate vector magnitude}

\item{duration}{the duration of processing}

\item{is_Message}{print update as a message?}
}
\description{
Print updates to console
}
\keyword{internal}
