% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_walkrun.R
\name{plot_walkrun}
\alias{plot_walkrun}
\title{Plot behaviors for visualizing a walk/run cut-point}
\usage{
plot_walkrun(object, activity_var, walkrun_activities, cpVar, MET_var, ...)
}
\arguments{
\item{object}{the \code{TwoRegression} object to plot from}

\item{activity_var}{character scalar. The name of the variable that indicates
the activity being performed}

\item{walkrun_activities}{character vector giving values of
\code{activity_var} that are walk/run activities}

\item{cpVar}{character scalar. The name of the variable on which the
cut-point is based}

\item{MET_var}{character scalar. The name of the variable containing energy
expenditure values, in metabolic equivalents}

\item{...}{additional arguments passed to \code{\link{plotCP}}}
}
\description{
Plot behaviors for visualizing a walk/run cut-point
}
\keyword{internal}
