% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2rm_helpers.R
\name{plotCP}
\alias{plotCP}
\title{Add a cut-point to a plot}
\usage{
plotCP(fig, cutPoint, x = NULL, y = NULL)
}
\arguments{
\item{fig}{A partially-assembled figure}

\item{cutPoint}{The threshold value to insert into the figure}

\item{x}{numeric scalar giving x coordinate for label placement}

\item{y}{numeric scalar giving y coordinate for label placement}
}
\description{
Add a cut-point to a plot
}
\keyword{internal}
