% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_TwoRegression.R
\name{predict.TwoRegression}
\alias{predict.TwoRegression}
\title{Predict metabolic equivalents from a TwoRegression object}
\usage{
\method{predict}{TwoRegression}(object, newdata, ...)
}
\arguments{
\item{object}{the TwoRegression object}

\item{newdata}{the data on which to predict metabolic equivalents}

\item{...}{further arguments passed to or from other methods}
}
\value{
A two-column data frame giving the activity classification
  (sedentary, walk/run, or intermittent activity) and the corresponding
  metabolic equivalent prediction
}
\description{
Predict metabolic equivalents from a TwoRegression object
}
\examples{
data(all_data, package = "TwoRegression")
newdata <- all_data

predict(Algorithms$Hip$Hibbing18_Hip_A1, newdata)
}
