% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0c_data_doc.R
\docType{data}
\name{raw_for_cv}
\alias{raw_for_cv}
\title{Primary accelerometer data to calculate coefficient of variation per 10-s}
\format{
A data frame with 299 rows and 2 variables:
\describe{
  \item{Block}{A vestigial variable synonymous with row number}
  \item{ENMO}{Euclidian Norm Minus One, in milli-g}
}
}
\usage{
raw_for_cv
}
\description{
A partially-processed primary accelerometer dataset ready to calculate the
coefficient of variation per 10-s
}
\keyword{datasets}
