% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_TwoRegression.R
\name{summary.TwoRegression}
\alias{summary.TwoRegression}
\title{Summary method for TwoRegression objects}
\usage{
\method{summary}{TwoRegression}(object, ...)
}
\arguments{
\item{object}{The TwoRegression object to summarize}

\item{...}{Arguments passed to \code{\link{DualCP_LOSO}}}
}
\value{
and object of class \code{summary.TwoRegression}, containing
  information about the cut-points and diagnostics, regression model
  performance, and overall algorithm performance during
  leave-one-participant-out-cross-validation.
}
\description{
Summary method for TwoRegression objects
}
\examples{

}
