% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates&Times.R, R/File_Reading_Functions.R,
%   R/File_Reading_Helper_Functions.R, R/Sensor_Functions.R,
%   R/TwoRegression-deprecated.R
\name{get_minute}
\alias{get_minute}
\alias{get_day_of_year}
\alias{read_AG_raw}
\alias{read_IMU}
\alias{check_columns}
\alias{check_second}
\alias{get_raw_file_meta}
\alias{get_imu_file_meta}
\alias{AG_collapse}
\alias{imu_collapse}
\alias{get_VM}
\alias{imu_filter_gyroscope}
\alias{classify_magnetometer}
\alias{TwoRegression-deprecated}
\title{Deprecated functions in package \pkg{TwoRegression}.}
\usage{
get_minute(timestamp, format = "\%Y-\%m-\%d \%H:\%M:\%S", rational = FALSE)

get_day_of_year(timestamp, format = "\%Y-\%m-\%d \%H:\%M:\%S")

read_AG_raw(file, output_window_secs = 1, verbose = FALSE)

read_IMU(file, output_window_secs = 1, verbose = FALSE)

check_columns(file)

check_second(AG)

get_raw_file_meta(file)

get_imu_file_meta(file, output_window_secs)

AG_collapse(AG, output_window_secs, samp_freq)

imu_collapse(AG, block_size, verbose = FALSE)

get_VM(triaxial, verbose = FALSE)

imu_filter_gyroscope(AG, samp_rate, verbose = FALSE)

classify_magnetometer(x = "Magnetometer X", y = "Magnetometer Y",
  z = "Magnetometer Z", orientation = "vertical")
}
\description{
These are functions that are deprecated because they have been
  moved elsewhere, primarily to the AGread package. Use
  \code{help("-deprecated")} to access documentation for deprecated
  functions.
}
\section{\code{get_minute}}{

For \code{get_minute}, use \code{\link[AGread]{get_minute}}
}

\section{\code{get_day_of_year}}{

For \code{get_day_of_year} use \code{\link[AGread]{get_day_of_year}}
}

\section{\code{read_AG_raw}}{

For \code{read_AG_raw}, use \code{\link[AGread]{read_AG_raw}}
}

\section{\code{read_IMU}}{

For \code{read_IMU}, use \code{\link[AGread]{read_AG_IMU}}
}

\section{\code{check_columns}}{

For \code{check_columns}, use \code{\link[AGread]{check_columns}}
}

\section{\code{check_second}}{

For \code{check_second}, use \code{\link[AGread]{check_second}}
}

\section{\code{get_raw_file_meta}}{

For \code{get_raw_file_meta}, use \code{\link[AGread]{get_raw_file_meta}}
}

\section{\code{get_imu_file_meta}}{

For \code{get_imu_file_meta}, use \code{\link[AGread]{get_imu_file_meta}}
}

\section{\code{AG_collapse}}{

For \code{AG_collapse}, use \code{\link[AGread]{AG_collapse}}
}

\section{\code{imu_collapse}}{

For \code{imu_collapse}, use \code{\link[AGread]{imu_collapse}}
}

\section{\code{get_VM}}{

For \code{get_VM}, use \code{\link[AGread]{get_VM}}
}

\section{\code{imu_filter_gyroscope}}{

For \code{imu_filter_gyroscope}, use \code{\link[AGread]{imu_filter_gyroscope}}
}

\section{\code{classify_magnetometer}}{

For \code{classify_magnetometer}, use \code{\link[AGread]{classify_magnetometer}}
}

\keyword{internal}
