% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoRegression.R
\docType{package}
\name{TwoRegression}
\alias{TwoRegression}
\alias{TwoRegression-package}
\title{TwoRegression: Process accelerometer data with common two-regression algorithms}
\description{
The critical functions of the TwoRegression package are:
}
\section{TwoRegression functions}{

I updated this file to see if it makes a difference.
}

